chore(dogstatsd): re-organize and clean up DSD codec
